Fixed some formatting.  Using continuations to properly leave javascript context now.  Just need to manage time now.
